+++ api/controllers/material.controller.js
@@ -11,6 +11,10 @@ const materialController = {
     }
   },
   createMaterials: async (req, res) => {
+    const { client } = req;
+    if (client.role !== 'admin') {
+      return res.status(403).json({ message: "Accès refusé. Vous n'êtes pas un administrateur." });
+    }
     try {
       const { name, value } = req.body;
       const material = await Material.create({ name, value });
@@ -21,6 +25,10 @@ const materialController = {
     }
   },
   updateMaterials: async (req, res) => {
+    const { client } = req;
+    if (client.role !== 'admin') {
+      return res.status(403).json({ message: "Accès refusé. Vous n'êtes pas un administrateur." });
+    }
     try {
       const { materialId } = req.params;
       const { name, value } = req.body;
@@ -43,6 +51,10 @@ const materialController = {
   },
 
   deleteMaterial: async (req, res) => {
+    const { client } = req;
+    if (client.role !== 'admin') {
+      return res.status(403).json({ message: "Accès refusé. Vous n'êtes pas un administrateur." });
+    }
     try {
       const { materialId } = req.params;
       const deleteMaterial = await Material.findByIdAndDelete(materialId);
