+++ api/controllers/productsController/product.controller.js
@@ -34,8 +34,10 @@ const productController = {
   }, 
  
   createProduct: async (req, res) => {
-    console.log('dans le controller createProduct');
-    console.log('req :',req.body);
+    const { client } = req;
+    if (client.role !== 'admin') {
+      return res.status(403).json({ message: "Accès refusé. Vous n'êtes pas un administrateur." });
+    }
     try {  
       const product = await Product.create(req.body); 
       res.status(201).json(product); 
@@ -45,6 +47,10 @@ const productController = {
   },
  
   updateProduct: async (req, res) => {
+    const { client } = req;
+    if (client.role !== 'admin') {
+      return res.status(403).json({ message: "Accès refusé. Vous n'êtes pas un administrateur." });
+    }
     try {
       const { productId } = req.params;
       const updateFields = req.body;
@@ -84,6 +90,10 @@ const productController = {
   },
 
   deleteProduct: async (req, res) => {
+    const { client } = req;
+    if (client.role !== 'admin') {
+      return res.status(403).json({ message: "Accès refusé. Vous n'êtes pas un administrateur." });
+    }
     try {
       const { productId } = req.params;
       await Product.findByIdAndDelete(productId);
